+++ app/src/androidTest/java/co/yml/ytag/app/ApplicationTest.kt
@@ -1,4 +1,4 @@
-package co.yml.ytag
+package co.yml.ytag.app
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.platform.app.InstrumentationRegistry
