+++ app/src/androidTest/java/co/yml/ytag/app/ui/MainNavigationKtTest.kt
@@ -1,9 +1,9 @@
-package co.yml.ytag.ui
+package co.yml.ytag.app.ui
 
 import androidx.compose.ui.test.assertIsDisplayed
 import androidx.compose.ui.test.junit4.createAndroidComposeRule
 import androidx.compose.ui.test.onNodeWithTag
-import co.yml.ytag.MainActivity
+import co.yml.ytag.app.MainActivity
 import dagger.hilt.android.testing.HiltAndroidRule
 import dagger.hilt.android.testing.HiltAndroidTest
 import org.junit.Rule
