+++ app/src/main/java/co/yml/ytag/app/YTag.kt
@@ -1,4 +1,4 @@
-package co.yml.ytag
+package co.yml.ytag.app
 
 import android.app.Application
 import dagger.hilt.android.HiltAndroidApp
